[root cause  ]:sync code
[changes     ]:sync code
[side effects]:none
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]:ums9230
[test case]:camera
[reviewers]:alec.zhu

+++ cam_sys/inc/cam_dump_node.h
@@ -54,7 +54,7 @@ enum cam_dump_node_id {
 };
 
 struct cam_dump_msg {
-	ssize_t size;
+	uint32_t size;
 	uint32_t align_w;
 	uint32_t align_h;
 	uint32_t offset;
